+++ CHANGELOG.md
@@ -7,11 +7,9 @@ This changelog format is largely based on [Keep A Changelog](https://keepachange
 
 #### 💥 Breaking Changes
 
-- No longer defaults package manager to pnpm if no flag is passed
-
 #### 🚀 New Features & Enhancements
 
-- Support setting a package manager in the config file
+- Detect package manager based on lockfile
 
 #### 🐛 Bug Fixes